FIX: バリテーション処理の修正
