+++ api/controller/user_controller.go
@@ -5,7 +5,9 @@ import (
 	"github.com/soramar/CBM_api/api/repository"
 	"github.com/gin-gonic/gin"
 	"github.com/soramar/CBM_api/model/schema"
+	"github.com/go-playground/validator/v10"
 	"golang.org/x/crypto/bcrypt"
+	"fmt"
 )
 
 type UserResponse struct {
@@ -33,9 +35,48 @@ func GetUsers(context *gin.Context) {
 
 func Register(c *gin.Context) {
 	var user schema.User
-	err := c.BindJSON(&user)
-	if err != nil {
+
+	if err := c.BindJSON(&user); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	validate := validator.New()
+
+	if err := validate.Struct(user); err != nil {
+		var errorMessages []string
+
+		for _, err := range err.(validator.ValidationErrors) {
+			var errorMessage string
+
+			fieldName := err.Field()
+			typ := err.Tag()
+			param := err.Param()
+
+			switch fieldName {
+				case "Name":
+					errorMessage = "名前は必須です"
+				case "Email":
+					switch typ {
+					case "required":
+						errorMessage = "メールアドレスは必須です"
+					case "email":
+						errorMessage = "メールアドレスのフォーマットで登録してください"
+					}
+				case "Password":
+					switch typ {
+					case "required":
+						errorMessage = "パスワードは必須です"
+					case "min":
+						errorMessage = fmt.Sprintf("パスワードは%s文字以上で登録してください", param)
+					}
+				case "Role":
+					errorMessage = "権限は必須です"
+			}
+			errorMessages = append(errorMessages, errorMessage)
+		}
+
+		c.JSON(http.StatusBadRequest, gin.H{"validation_error": errorMessages})
 		return
 	}
 
@@ -47,8 +88,7 @@ func Register(c *gin.Context) {
 
 	user.Password = hashedPassword
 
-	err = repository.CreateUser(&user)
-	if err != nil {
+	if err := repository.CreateUser(&user); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
